backend/requirements.txt
boto3==1.34.144
botocore==1.34.144
celery==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-storages==1.14.4
ffmpeg-python==0.2.0
psycopg==3.2.1
psycopg-binary==3.2.1
pycurl==7.45.3
python-dotenv==1.0.1
ultralytics==8.2.64
